Save of Room Reservation is implemented. Need to test. In utility new converter is added for double also.

+++ Project/AutoTourism/Source Code/AutoTourism/Lodge/AutoTourism.Lodge.Facade/RoomReservation/Dto.cs
@@ -11,10 +11,10 @@ namespace AutoTourism.Lodge.Facade.RoomReservation
         public Int64 Id { get; set; }
         public DateTime BookingDate { get; set; }
         public DateTime BookingFrom { get; set; }
-        public Int32 NoOfDays { get; set; }
-        public Int32 NoOfPersons { get; set; }
-        public Int32 NoOfRooms { get; set; }
-        public Decimal Advance { get; set; }
+        public Int16 NoOfDays { get; set; }
+        public Int16 NoOfPersons { get; set; }
+        public Int16 NoOfRooms { get; set; }
+        public Double Advance { get; set; }
         public Int64 BookingStatusId { get; set; }
 
         public List<LodgeConfigurationFacade.Room.Dto> RoomList { get; set; }
